user permission

+++ commands/deleteJob.js
@@ -7,26 +7,31 @@ module.exports = {
     description: "delete job",
     async execute(msg, args) {
         const embed = new MessageEmbed().setColor("#add79b");
-		let job;
-        let num = Number(args);
+        if(msg.member.roles.cache.has("909049029734834226")) {
+            let job;
+            let num = Number(args);
 
-        if(!Number.isInteger(num)) {
-            embed.setDescription("Invalid Id");
-            msg.reply({ embeds: [embed] });
-            return;
-        } 
+            if(!Number.isInteger(num)) {
+                embed.setDescription("Invalid Id");
+                msg.reply({ embeds: [embed] });
+                return;
+            } 
 
-		job = await jobModel.findOne({serverId: msg.guild.id, jobId: num })
-        if(!job) {
-            embed.setDescription("Invalid Id");
+            job = await jobModel.findOne({serverId: msg.guild.id, jobId: num })
+            if(!job) {
+                embed.setDescription("Invalid Id");
+                msg.reply({ embeds: [embed] });
+                return;
+            }
+
+            await job.deleteOne({});
+            embed.setDescription(`Job ${job.jobName} has been deleted`);
             msg.reply({ embeds: [embed] });
-            return;
+
+            updateId(msg.guild.id);
+        } else {
+            embed.setDescription("You can't send this command because you don't have the right permissions");
+            msg.reply({ embeds: [embed] });
         }
-
-        await job.deleteOne({});
-        embed.setDescription(`Job ${job.jobName} has been deleted`);
-        msg.reply({ embeds: [embed] });
-
-        updateId(msg.guild.id);
     }
 };
